refactor navigator
