+++ screens/WalletNavigator/WalletNavigator.tsx
