+++ screens/WalletNavigator/screens/WalletMnemonicCreate.tsx
