+++ screens/WalletNavigator/screens/WalletOnboarding.tsx
